Organize the code
Moving ini related functions to config.h
